requirements.txt
transformers~=4.25.1
torch~=1.13.0
streamlit~=1.15.2
Pillow~=9.3.0